requirements.txt
datasets
pandas
numpy
openai
python-dotenv
matplotlib
